+++ app/src/main/java/org/bibichan/union/player/ui/player/PlayerViewModelFactory.kt
@@ -1,16 +1,18 @@
 package org.bibichan.union.player.ui.player
 
+import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import org.bibichan.union.player.MusicPlayer
 
 class PlayerViewModelFactory(
-    private val musicPlayer: MusicPlayer
+    private val musicPlayer: MusicPlayer,
+    private val context: Context? = null
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(PlayerViewModel::class.java)) {
             @Suppress("UNCHECKED_CAST")
-            return PlayerViewModel(musicPlayer) as T
+            return PlayerViewModel(musicPlayer, context) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class: ${modelClass.name}")
     }